Move a bunch of code from host.rs to {auth,lifecycle,data_helper}.rs
